refactor: implement Phase 5 - Presentation Layer Integration
- Update all pages to import from domain layer instead of legacy hooks:
  - WorksList.tsx, WorkForm.tsx, Categories.tsx, Dashboard.tsx

- Update components to use core/types:
  - VideoUploader.tsx, MediaOrderManager.tsx
  - CaptionEditor.tsx (also updated to use domain hook)

- Remove legacy adapter folders:
  - src/hooks/ (was re-exporting from domain)
  - src/types/ (was re-exporting from core/types)

- Update test imports to use new module paths

All 194 tests passing, build successful, lint clean.
