+++ admin/src/__tests__/hooks/useWorks.test.tsx
@@ -8,8 +8,8 @@ import {
   useCreateWork,
   useUpdateWork,
   useDeleteWork,
-  worksKeys,
-} from '../../hooks/useWorks';
+} from '../../domain';
+import { worksCacheKeys as worksKeys } from '../../data/repository';
 import {
   getWorks,
   getWork,
