--- admin/src/types/index.ts
@@ -1,3 +1,0 @@
-// Re-export all types from core/types for backward compatibility
-// TODO: Update all imports to use '@/core/types' directly
-export * from '../core/types';
